requirements.txt
langchain
langchain_groq
sentence-transformers
faiss-cpu
PyMuPDF
python-dotenv
fastapi
uvicorn
